+++ anybox/buildbot/openerp/subfactories/__init__.py
@@ -8,6 +8,7 @@ buildout_download = dict(standalone=download.standalone_buildout,
                          bzr=download.bzr_buildout,
                          hg=download.hg_buildout,
                          git=download.git_buildout,
+                         archive=download.archive_buildout,
                          )
 
 db_handling = dict(simple_create=db.simple_create,